seyba

+++ src/main/java/org/psychohelp/psychohelp/repository/SeanceRepository.java
@@ -16,5 +16,5 @@ public interface SeanceRepository extends JpaRepository<Seance, Long> {
     List<Seance> findByStatut(StatutRdvEnum statut);
 
     @Query("SELECT s FROM Seance s JOIN s.creneau c WHERE c.psychologue.id = :psyId")
-    List<Seance> findByPsyId(@Param("psyId") Long psyId);
+    List<Seance> findByPsyId(@Param("psyId") int psyId);
 }
